date validator as a mixin
